Revert "Generate MySQL root password randomly"
