+++ db/Makefile
@@ -1,10 +1,21 @@
+root_password="DtWb3q2WcG0tc2fFgXi5sGx86XL7RT8g"
 all: mysql
 
 mysql:
-	./setup_mysql.sh
-	sudo apt-get -y install python3-pip
+	echo PURGE | sudo debconf-communicate mysql-server
+	echo PURGE | sudo debconf-communicate mysql-server-5.5
+	echo "mysql-server mysql-server/root_password password ${root_password}" | sudo debconf-set-selections
+	echo "mysql-server mysql-server/root_password_again password ${root_password}" | sudo debconf-set-selections
+	sudo apt-get update
+	sudo apt-get -y install mysql-server python3-pip
 	sudo pip3 install PyMySQL
 	sudo pip3 install tabulate
+	mysql -uroot -p${root_password} < create_bankdb.sql
+	./get_pubkey.sh
+	python3 reg_pubkey.py
+	python3 reg_admin.py
+	echo PURGE | sudo debconf-communicate mysql-server
+	echo PURGE | sudo debconf-communicate mysql-server-5.5
 
 .PHONY: all mysql
