--- db/gen_root_pw.py
@@ -1,7 +1,0 @@
-import utils
-
-if __name__ == "__main__":
-    pw = utils.gen_rand_pw(32)
-    with open("root.pw", "wt") as f:
-        f.write(pw)
-
