+++ db/utils.py
@@ -3,7 +3,7 @@ import pymysql.cursors
 class bankDB:
     def __init__(self, host='localhost',
                        user='root',
-                       password = open("root.pw","rt").read(),
+                       password='DtWb3q2WcG0tc2fFgXi5sGx86XL7RT8g',
                        db='bankDB'):
         '''
             localhost, root, ~
@@ -230,12 +230,3 @@ class bankDB:
     # Disconnect
     def close_conn(self):
         self.conn.close()
-
-def gen_rand_pw(N):
-    import os
-    import random
-    import string
-    random.seed(os.urandom(64))
-    return ''.join(random.choice(string.ascii_uppercase + \
-                   string.ascii_lowercase + string.digits) \
-                   for _ in range(N))
